export default from entry auto wraps with Motion.entry
